add session power manager

+++ src/gql-game/advert/resolver.ts
@@ -19,6 +19,8 @@ export class Advert {
 
 	static Query: any = {
 		getAdverts(_, __, context): Promise<Array<IAdvertModel>> {
+			if(!context.user) return null;
+
 			var uploadFile = new UploadFile();
 			let promise = new Promise<Array<IAdvertModel>>((resolve, reject) => {
 				AdvertSchema.find().then((res) => {
@@ -29,6 +31,8 @@ export class Advert {
 		},
 
 		getAdvertById(_, { id }, context): Promise<IAdvertModel> {
+			if(!context.user) return null;
+
 			let promise = new Promise<IAdvertModel>((resolve, reject) => {
 				AdvertSchema.findById(id).then((res) => {
 					resolve(res);
@@ -38,16 +42,22 @@ export class Advert {
 		},
 
 		getAdvertPage(_, { pageIndex = 1, pageSize = 10, advert }, context) {
+			if(!context.user) return null;
+
 			var userInfo = AdvertSchema.find(advert).skip((pageIndex - 1) * pageSize).limit(pageSize);
 			return userInfo;
 		},
 
 		getAdvertCount(_, { advert }, context) {
+			if(!context.user) return null;
+
 			var count = AdvertSchema.count(advert);
 			return count;
 		},
 
 		getAdvertWhere(_, { advert }, context) {
+			if(!context.user) return null;
+
 			var users = AdvertSchema.find(advert);
 			return users;
 		}
@@ -55,6 +65,8 @@ export class Advert {
 
 	static Mutation: any = {
 		saveAdvert(_, { advert }, context) {
+			if(!context.user) return null;
+
 			if (advert.id) {
 				return new Promise<IAdvertModel>((resolve, reject) => {
 					AdvertSchema.findByIdAndUpdate(advert.id, advert, (err, res) => {
@@ -67,6 +79,8 @@ export class Advert {
 		},
 
 		deleteAdvert(_, { id }, context): Promise<Boolean> {
+			if(!context.user) return null;
+
 			let promise = new Promise<Boolean>((resolve, reject) => {
 				AdvertSchema.findByIdAndRemove(id, (err, res) => {
 					new UploadFile().deleteImg(res.imageId);
