+++ src/gql-system/power/resolver.ts
@@ -11,6 +11,8 @@ export class Power {
     static Query: any = {
 
         getPowers(_, __, context): Promise<Array<IPowerModel>> {
+            if(!context.user) return null;
+
             let promise = new Promise<Array<IPowerModel>>((resolve, reject) => {
                 PowerSchema.find().then(res => {
                     resolve(res);
@@ -20,6 +22,8 @@ export class Power {
         },
 
         getPowerById(_, { id }, context): Promise<IPowerModel> {
+            if(!context.user) return null;
+
             let promise = new Promise<IPowerModel>((resolve, reject) => {
                 PowerSchema.findById(id).then(res => {
                     resolve(res);
@@ -28,18 +32,22 @@ export class Power {
             return promise;
         },
 
-        getPowerPage(_, { pageIndex = 1, pageSize = 10, power }, context):
-            DocumentQuery<Array<IPowerModel>, IPowerModel> {
-            var list = PowerSchema.find(power).skip((pageIndex - 1) * pageSize).limit(pageSize)
+        getPowerPage(_, { pageIndex = 1, pageSize = 10, power }, context) {
+            if(!context.user) return null;
+            var skip=(pageIndex - 1) * pageSize;
+            var list = PowerSchema.find(power).skip(skip).limit(pageSize)
             return list;
         },
 
         getPowerCount(_, { power }, context) {
+            if(!context.user) return 0;
             var count = PowerSchema.count(power);
             return count;
         },
 
         getPowerWhere(_, { power }, context) {
+            if(!context.user) return null;
+            
             var users = PowerSchema.find(power);
             return users;
         },
@@ -47,6 +55,8 @@ export class Power {
 
     static Mutation: any = {
         savePower(_, { power }, context) {
+            if(!context.user) return null;
+            
             if (power.id) {
                 return new Promise<IPowerModel>((resolve, reject) => {
                     PowerSchema.findByIdAndUpdate(power.id, power, (err, res) => {
@@ -60,6 +70,8 @@ export class Power {
         },
 
         deletePower(_, { id }, context): Promise<Boolean> {
+            if(!context.user) return null;
+            
             let promise = new Promise<Boolean>((resolve, reject) => {
                 PowerSchema.findByIdAndRemove(id, (err, res) => {
                     resolve(res != null)
@@ -67,10 +79,14 @@ export class Power {
             });
             return promise;
         },
-        addAllPower(_, { power }, context): MongoosePromise<Array<IPowerModel>> {
+        addAllPower(_, { power }, context){
+            if(!context.user) return null;
+            
             return PowerSchema.create(power);
         },
         delAllPower(_, { power }, context): Promise<Boolean> {
+            if(!context.user) return null;
+            
             let promise = new Promise<Boolean>((resolve, reject) => {
                 if (!power) resolve(false);
                 PowerSchema.find(power).remove((err, res) => {
